+++ clvm_tools/binutils.ts
@@ -31,41 +31,77 @@ function isPrintable(s: string){
   return regex.test(s);
 }
 
+// In order to reduce stack memory consumed, I made `assemble_from_ir` fully flatten from the previous recursive function callstack.
 export function assemble_from_ir(ir_sexp: SExp): SExp {
-  let keyword = ir_as_symbol(ir_sexp);
-  if(keyword){
-    if(keyword[0] === "#"){
-      keyword = keyword.substring(1);
+  const input_stack: Array<[number, SExp]> = []; // [depth, SExp]
+  const return_value_stack: Array<[number, SExp]> = []; // [depth, SExp]
+  let last_return_value: SExp|undefined;
+  
+  input_stack.push([0, ir_sexp]);
+  while(input_stack.length || return_value_stack.length){
+    while(return_value_stack.length >= 2){
+      if(return_value_stack[return_value_stack.length-1][0] !== return_value_stack[return_value_stack.length-2][0]){
+        break;
+      }
+      const sexp_1 = return_value_stack.pop() as [number, SExp];
+      const sexp_2 = return_value_stack.pop() as [number, SExp];
+      last_return_value = sexp_1[1].cons(sexp_2[1]);
+      return_value_stack.push([sexp_1[0]-1, last_return_value]);
     }
-    const atom = KEYWORD_TO_ATOM[keyword as keyof typeof KEYWORD_TO_ATOM];
-    if(atom){
-      return SExp.to(h(atom));
+    if(!input_stack.length){
+      if(return_value_stack.length === 1){
+        break;
+      }
+      continue;
     }
-    else{
-      return ir_val(ir_sexp);
+    
+    const depth_and_sexp = (input_stack.pop() as [number, SExp]);
+    ir_sexp = depth_and_sexp[1];
+    let keyword = ir_as_symbol(ir_sexp);
+    if(keyword){
+      if(keyword[0] === "#"){
+        keyword = keyword.substring(1);
+      }
+      const atom = KEYWORD_TO_ATOM[keyword as keyof typeof KEYWORD_TO_ATOM];
+      if(atom){
+        last_return_value = SExp.to(h(atom));
+        return_value_stack.push([depth_and_sexp[0], last_return_value]);
+        continue;
+      }
+      else{
+        last_return_value = ir_val(ir_sexp);
+        return_value_stack.push([depth_and_sexp[0], last_return_value]);
+        continue;
+      }
+      // Original code raises an Error, which never reaches.
+      // throw new SyntaxError(`can't parse ${keyrowd} at ${ir_sexp._offset}`);
     }
-    // Original code raises an Error, which never reaches.
-    // throw new SyntaxError(`can't parse ${keyrowd} at ${ir_sexp._offset}`);
+  
+    if(!ir_listp(ir_sexp)){
+      last_return_value = ir_val(ir_sexp);
+      return_value_stack.push([depth_and_sexp[0], last_return_value]);
+      continue;
+    }
+  
+    if(ir_nullp(ir_sexp)){
+      last_return_value = SExp.to([]);
+      return_value_stack.push([depth_and_sexp[0], last_return_value]);
+      continue;
+    }
+  
+    // handle "q"
+    const first = ir_first(ir_sexp);
+    keyword = ir_as_symbol(first);
+    if(keyword === "q"){
+      // pass;
+    }
+  
+    const depth = depth_and_sexp[0] + 1;
+    input_stack.push([depth, first]);
+    input_stack.push([depth, ir_rest(ir_sexp)]);
   }
   
-  if(!ir_listp(ir_sexp)){
-    return ir_val(ir_sexp);
-  }
-  
-  if(ir_nullp(ir_sexp)){
-    return SExp.to([]);
-  }
-  
-  // handle "q"
-  const first = ir_first(ir_sexp);
-  keyword = ir_as_symbol(first);
-  if(keyword === "q"){
-    // pass;
-  }
-  
-  const sexp_1 = assemble_from_ir(first);
-  const sexp_2 = assemble_from_ir(ir_rest(ir_sexp));
-  return sexp_1.cons(sexp_2);
+  return last_return_value as SExp;
 }
 
 export function type_for_atom(atom: Bytes): number {