+++ platform/performance.ts
@@ -1,5 +1,5 @@
 export function now(){
   // eslint-disable-next-line @typescript-eslint/no-var-requires
   const {performance} = require("perf_hooks");
-  return performance.now();
+  return performance.now() / 1000.0;
 }